made record.errors an object with arrays of errors for each field.
Need to clarify how to add non-field-specific custom validation in docs.

Several validation methods need verification/testing/fixing, as noted in todo.txt

Validation methods need to be able to accept an array of colnames or a single string

Cleaned up isValid so it knows which validation to call automatically when called manually.